Retype mbyte's anonymous enums and give MB_MAXCHAR one home
B15-8's second half, before a line of behaviour is rewritten: c2rust
types an anonymous C enum from what the C compiler picked, not from what
the consumers read, so every use site casts it back. Retyping is only a
cleanup if the casts go with it, which is why this is its own commit and
why **clippy is unchanged at 3,818** across it.

- **`C2Rust_Unnamed_19` -> `EncProps` (`c_int`)** — the `ENC_*` bit set
  `enc_canon_props()` answers with. `c_int` is the type of
  `enc_canon_table`'s `prop` field and of everything compared against it;
  **130 casts** go, 122 of them in the table itself.
- **`ConvFlags` -> `c_int`** — `vimconv_T::vc_type`'s own type. B15-2
  consolidated the six `CONV_*` at the transpiled `c_uint` and left a
  note that the retype belonged to the slice deleting the casts. This is
  that slice: **55 sites across 14 files**, only a fifth of them in
  mbyte.
- **`C2Rust_Unnamed_23` -> `nl_item`** — `CODESET` is an argument to
  `nl_langinfo`, which takes exactly that.
- **`C2Rust_Unnamed`** — the `<ctype.h>` class bits. Four of the five
  (`_ISpunct`, `_IScntrl`, `_ISgraph`, `_ISalpha`) had no user at all;
  `_ISalnum` becomes the bare `c_uint` the six other copies in the tree
  already are, and its use site loses a `c_int`-to-`c_ushort`-to-`c_int`
  round trip.

**`MB_MAXCHAR` had four definitions** — mbyte's `c_uint`, `match.rs`'s
and `message/outtrans.rs`'s private `usize` copies — and moves to
`types/mbyte.rs` beside `MB_MAXBYTES`, as `usize`, which is what every
one of its uses is (an array length). Both consumers that imported it
from mbyte now take it from `types`, and the three `as usize` casts go.

Gates: utfsweep IDENTICAL on all four artifacts; `just unittest` 771/778;
`just cargo-test` 602. `just refresh` with **no `--allow-growth`**.

+++ crates/nvim/src/nvim/eval/collect.rs
@@ -653,9 +653,7 @@ pub unsafe fn var_item_copy(
         let mut ret = OK;
         match (*from).v_type {
             VAR_STRING => {
-                if conv.is_null()
-                    || (*conv).vc_type == CONV_NONE as c_int
-                    || (*from).vval.v_string.is_null()
+                if conv.is_null() || (*conv).vc_type == CONV_NONE || (*from).vval.v_string.is_null()
                 {
                     tv_copy(from, to);
                 } else {
