+++ crates/nvim/src/nvim/spellfile/aff.rs
@@ -333,7 +333,7 @@ pub unsafe fn spell_read_aff(spin: *mut spellinfo_T, fname: *mut c_char) -> *mut
 
             xfree(pc.cast());
             pc = core::ptr::null_mut();
-            let line = if (*spin).si_conv.vc_type != CONV_NONE as c_int {
+            let line = if (*spin).si_conv.vc_type != CONV_NONE {
                 pc = string_convert(
                     &raw mut (*spin).si_conv,
                     rline.as_mut_ptr(),
