+++ crates/nvim/src/nvim/viml/parser/parser.rs
@@ -13,7 +13,7 @@
 //! reports as soon as the parser is driven end to end. Reborrows here are
 //! narrowed to the one collection being touched.
 
-use core::ffi::{c_char, c_int, c_void};
+use core::ffi::{c_char, c_void};
 use core::ptr;
 
 use crate::src::nvim::kvec::InitVec;
@@ -49,7 +49,7 @@ pub const PARSER_STATE_INIT: ParserState = ParserState {
             init_array: [EMPTY_LINE; 4],
         },
         conv: vimconv_T {
-            vc_type: CONV_NONE as c_int,
+            vc_type: CONV_NONE,
             vc_factor: 1,
             vc_fd: ptr::null_mut(),
             vc_fail: false,
@@ -146,7 +146,7 @@ pub unsafe fn viml_parser_get_remaining_line(pstate: *mut ParserState) -> Option
         let mut fresh = EMPTY_LINE;
         let get_line = (*reader).get_line.expect("parser has no line getter");
         get_line((*reader).cookie, &raw mut fresh);
-        if (*reader).conv.vc_type != CONV_NONE as c_int && fresh.size != 0 {
+        if (*reader).conv.vc_type != CONV_NONE && fresh.size != 0 {
             let mut converted = ParserLine {
                 data: ptr::null(),
                 size: fresh.size,
